Refactored includes and CMake to enable building RMM as a submodule
