+++ .gitignore
@@ -34,3 +34,6 @@ python/librmm_cffi/librmm_cffi.py
 .project
 .cproject
 .settings
+
+# cnmem copy
+include/rmm/detail/cnmem.h
