+++ src/memory.cpp
@@ -21,9 +21,10 @@
  *
  */
 
-#include "rmm.h"
-#include "memory_manager.h"
-#include "memory.hpp"
+#include "rmm/rmm.h"
+#include "rmm/detail/memory_manager.hpp"
+#include "rmm/memory.hpp"
+
 #include <fstream>
 #include <sstream>
 #include <cstddef>
